Create Hero vs. Monster
Learning goals:
- Static member variables and method
- Singleton design pattern
- Object-Oriented Programming (OOP)

Hero.java
This class creates a blueprint for making a singleton Hero object. The singleton hero object has certain attributes and can interact with Item and Monster objects. You will complete the methods in this file. For those methods that are already implemented, please understand them and do NOT change them.

Monster.java
This class creates a blueprint for making Monster objects. Each monster has certain attributes. Multiple monsters can be created in the program. It can interact with the Hero object. You will complete the methods in this file. For those methods that are already implemented, please understand them and do NOT change them.

Item.java	
This class creates a blueprint for making Item objects. Each item has certain attributes. Each Item object can interact with the Hero object. You will complete the methods in this file. For those methods that are already implemented, please understand them and do NOT change them.

Tower.java	
This class creates a blueprint for making Tower objects. Each Tower object contains multiple Monster objects and Item objects. The Hero object will take an adventure in the tower, fighting monsters and taking items. You will complete the methods in this file. For those methods that are already implemented, please understand them and do NOT change them.

PA6Tester.java 
This is a simple tester to run/test the methods you will be implementing. It does not cover all test cases for correct implementation. You might want to add more tests to this file. For those methods that are already implemented, please understand them and do NOT change them.